# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type QualityIssueSeverity = typeof QualityIssueSeverity[keyof typeof QualityIssueSeverity];


export const QualityIssueSeverity = {
  error: 'error',
  warning: 'warning',
  review: 'review',
} as const;

export interface QualityIssue {
  code: string;
  severity: QualityIssueSeverity;
  title: string;
  detail: string;
  rowNumbers: number[];
}

export interface PreviewRow {
  sourceRow: number;
  /** @nullable */
  systemId: string | null;
  /** @nullable */
  identifier: string | null;
  /** @nullable */
  nomenclature: string | null;
  /** @nullable */
  unit: string | null;
  /** @nullable */
  currentDglp: number | null;
  /** @nullable */
  currentEchs: number | null;
}

export type LegacyRowInputDepartmentsItem = {
  departmentIndex: number;
  /** @nullable */
  pvms?: string | null;
  /** @nullable */
  dglp?: number | null;
  /** @nullable */
  echs?: number | null;
};

export interface LegacyRowInput {
  sourceRow: number;
  /** @nullable */
  systemId?: string | null;
  /** @nullable */
  identifier?: string | null;
  /** @nullable */
  nomenclature?: string | null;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  previousPvmsMmf?: number | null;
  /** @nullable */
  currentPvmsMmf?: number | null;
  /** @nullable */
  previousDglpMmf?: number | null;
  /** @nullable */
  currentDglpMmf?: number | null;
  /** @nullable */
  previousEchsMmf?: number | null;
  /** @nullable */
  currentEchsMmf?: number | null;
  /** @nullable */
  lpr?: string | null;
  sourceValues: unknown[];
  departments: LegacyRowInputDepartmentsItem[];
}

export interface DepartmentInput {
  name: string;
  sourceColumnStart: number;
}

export interface ImportInput {
  sourceFileName: string;
  sourceFileHash: string;
  fileSize?: number;
  worksheetName: string;
  rowCount: number;
  columnCount: number;
  departmentCount: number;
  uniqueIdentifierCount: number;
  missingIdentifierCount: number;
  conflictGroupCount: number;
  warnings: number;
  errors: number;
  previewRows: PreviewRow[];
  issues?: QualityIssue[];
  legacyRows?: LegacyRowInput[];
  departments?: DepartmentInput[];
  /** @nullable */
  sourceObjectPath?: string | null;
  /** @nullable */
  sourceObjectContentType?: string | null;
}

export interface CommitImportInput {
  acknowledged?: boolean;
}

export type ImportSummaryStatus = typeof ImportSummaryStatus[keyof typeof ImportSummaryStatus];


export const ImportSummaryStatus = {
  review: 'review',
  committed: 'committed',
} as const;

export type ImportSummaryQuality = {
  warnings: number;
  errors: number;
  conflicts: number;
  missingIdentifiers: number;
};

export interface ImportSummary {
  id: string;
  sourceFileName: string;
  worksheetName: string;
  rowCount: number;
  columnCount?: number;
  departmentCount: number;
  status: ImportSummaryStatus;
  createdAt: string;
  /** @nullable */
  committedAt?: string | null;
  quality: ImportSummaryQuality;
}

export type ImportDetail = ImportSummary & {
  sourceFileHash: string;
  fileSize: number;
  uniqueIdentifierCount: number;
  previewRows: PreviewRow[];
  issues: QualityIssue[];
};

export interface Department {
  id: string;
  name: string;
  itemCount: number;
}

export interface CanonicalItem {
  id: string;
  canonicalId: string;
  /** @nullable */
  nomenclature: string | null;
  /** @nullable */
  unit: string | null;
  /** @nullable */
  pvms: string | null;
  /** @nullable */
  niv: string | null;
  status: string;
  legacyRecordCount: number;
  departmentCount: number;
}

export type LegacyItemDepartmentsItem = {
  name: string;
  /** @nullable */
  pvms: string | null;
  /** @nullable */
  dglp: number | null;
  /** @nullable */
  echs: number | null;
};

export interface LegacyItem {
  id: string;
  importId: string;
  sourceWorksheet: string;
  sourceRow: number;
  /** @nullable */
  systemId: string | null;
  /** @nullable */
  identifier: string | null;
  /** @nullable */
  nomenclature: string | null;
  /** @nullable */
  unit: string | null;
  /** @nullable */
  previousPvmsMmf: number | null;
  /** @nullable */
  currentPvmsMmf: number | null;
  /** @nullable */
  previousDglpMmf: number | null;
  /** @nullable */
  currentDglpMmf: number | null;
  /** @nullable */
  previousEchsMmf: number | null;
  /** @nullable */
  currentEchsMmf: number | null;
  /** @nullable */
  lpr: string | null;
  sourceValues: unknown[];
  departments: LegacyItemDepartmentsItem[];
}

export type CanonicalItemDetailVocabularyHistoryItem = {
  id: string;
  reviewType: string;
  status: string;
  title: string;
  detail: string;
};

export type CanonicalItemDetail = CanonicalItem & {
  legacyRecords: LegacyItem[];
  vocabularyHistory: CanonicalItemDetailVocabularyHistoryItem[];
};

export interface CanonicalItemPage {
  items: CanonicalItem[];
  page: number;
  pageSize: number;
  total: number;
}

export interface CreateCanonicalItemInput {
  /** @minItems 1 */
  legacyItemIds: string[];
  /** @nullable */
  nomenclature?: string | null;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  pvms?: string | null;
  /** @nullable */
  niv?: string | null;
}

export interface VocabularyReview {
  id: string;
  importId: string;
  reviewType: string;
  /** @nullable */
  identifier: string | null;
  title: string;
  detail: string;
  status: string;
  /** @nullable */
  decision?: string | null;
  /** @nullable */
  decisionNote?: string | null;
  candidateRecords: LegacyItem[];
}

export type VocabularyDecisionInputDecision = typeof VocabularyDecisionInputDecision[keyof typeof VocabularyDecisionInputDecision];


export const VocabularyDecisionInputDecision = {
  MERGE: 'MERGE',
  KEEP_SEPARATE: 'KEEP_SEPARATE',
  CORRECT: 'CORRECT',
  RETIRE: 'RETIRE',
  CREATE_CANONICAL: 'CREATE_CANONICAL',
  INVESTIGATE: 'INVESTIGATE',
} as const;

export interface VocabularyDecisionInput {
  decision: VocabularyDecisionInputDecision;
  note?: string;
  canonicalItemIds?: string[];
  sourceColumnStart?: number;
}

export type OverviewCycle = {
  name: string;
  status: string;
  deadline: string;
  daysRemaining: number;
};

export type OverviewBaseline = {
  rowCount: number;
  departmentCount: number;
  uniqueIdentifiers: number;
  currentImportStatus: string;
};

export type OverviewQuality = {
  conflicts: number;
  warnings: number;
  unresolved: number;
};

export type OverviewDepartments = {
  total: number;
  ready: number;
  inProgress: number;
  needsReview: number;
};

export type OverviewRecentActivityItemTone = typeof OverviewRecentActivityItemTone[keyof typeof OverviewRecentActivityItemTone];


export const OverviewRecentActivityItemTone = {
  neutral: 'neutral',
  success: 'success',
  warning: 'warning',
  critical: 'critical',
} as const;

export type OverviewRecentActivityItem = {
  id: string;
  action: string;
  detail: string;
  timestamp: string;
  tone: OverviewRecentActivityItemTone;
};

export interface Overview {
  cycle: OverviewCycle;
  baseline: OverviewBaseline;
  quality: OverviewQuality;
  departments: OverviewDepartments;
  recentActivity: OverviewRecentActivityItem[];
}

export type ListCanonicalItemsParams = {
search?: string;
status?: string;
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
pageSize?: number;
};

export type ListVocabularyReviewsParams = {
status?: ListVocabularyReviewsStatus;
};

export type ListVocabularyReviewsStatus = typeof ListVocabularyReviewsStatus[keyof typeof ListVocabularyReviewsStatus];


export const ListVocabularyReviewsStatus = {
  open: 'open',
  resolved: 'resolved',
  all: 'all',
} as const;