# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /canonical-items:
    get:
      operationId: listCanonicalItems
      tags: [vocabulary]
      summary: Search canonical vocabulary
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
      responses:
        "200":
          description: Canonical vocabulary page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanonicalItemPage"
    post:
      operationId: createCanonicalItem
      tags: [vocabulary]
      summary: Create a canonical item from immutable legacy records
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCanonicalItemInput"
      responses:
        "201":
          description: Canonical item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanonicalItemDetail"
        "409":
          description: A legacy record is already linked
  /canonical-items/{canonicalItemId}:
    get:
      operationId: getCanonicalItem
      tags: [vocabulary]
      summary: Get canonical item detail and lineage
      parameters:
        - $ref: "#/components/parameters/CanonicalItemId"
      responses:
        "200":
          description: Canonical item detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanonicalItemDetail"
        "404":
          description: Canonical item not found
  /vocabulary-reviews:
    get:
      operationId: listVocabularyReviews
      tags: [vocabulary]
      summary: List vocabulary review records
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, resolved, all]
      responses:
        "200":
          description: Vocabulary review queue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VocabularyReview"
  /vocabulary-reviews/{reviewId}/decision:
    post:
      operationId: decideVocabularyReview
      tags: [vocabulary]
      summary: Record a human vocabulary review decision
      parameters:
        - $ref: "#/components/parameters/ReviewId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VocabularyDecisionInput"
      responses:
        "200":
          description: Review decision recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VocabularyReview"
        "404":
          description: Review not found
        "409":
          description: Decision cannot be applied
components:
  parameters:
...
    CanonicalItemId:
      name: canonicalItemId
      in: path
      required: true
      schema:
        type: string
    ReviewId:
      name: reviewId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    CanonicalItem:
      type: object
      required: [id, canonicalId, nomenclature, unit, pvms, niv, status, legacyRecordCount, departmentCount]
      properties:
        id:
          type: string
        canonicalId:
          type: string
        nomenclature:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        pvms:
          type: ["string", "null"]
        niv:
          type: ["string", "null"]
        status:
          type: string
        legacyRecordCount:
          type: integer
        departmentCount:
          type: integer
    LegacyItem:
      type: object
      required: [id, importId, sourceWorksheet, sourceRow, systemId, identifier, nomenclature, unit, previousPvmsMmf, currentPvmsMmf, previousDglpMmf, currentDglpMmf, previousEchsMmf, currentEchsMmf, lpr, sourceValues, departments]
      properties:
        id:
          type: string
        importId:
          type: string
        sourceWorksheet:
          type: string
        sourceRow:
          type: integer
        systemId:
          type: ["string", "null"]
        identifier:
          type: ["string", "null"]
        nomenclature:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        previousPvmsMmf:
          type: ["number", "null"]
        currentPvmsMmf:
          type: ["number", "null"]
        previousDglpMmf:
          type: ["number", "null"]
        currentDglpMmf:
          type: ["number", "null"]
        previousEchsMmf:
          type: ["number", "null"]
        currentEchsMmf:
          type: ["number", "null"]
        lpr:
          type: ["string", "null"]
        sourceValues:
          type: array
          items: {}
        departments:
          type: array
          items:
            type: object
            required: [name, pvms, dglp, echs]
            properties:
              name:
                type: string
              pvms:
                type: ["string", "null"]
              dglp:
                type: ["number", "null"]
              echs:
                type: ["number", "null"]
    CanonicalItemDetail:
      allOf:
        - $ref: "#/components/schemas/CanonicalItem"
        - type: object
          required: [legacyRecords, vocabularyHistory]
          properties:
            legacyRecords:
              type: array
              items:
                $ref: "#/components/schemas/LegacyItem"
            vocabularyHistory:
              type: array
              items:
                type: object
                required: [id, reviewType, status, title, detail]
                properties:
                  id:
                    type: string
                  reviewType:
                    type: string
                  status:
                    type: string
                  title:
                    type: string
                  detail:
                    type: string
    CanonicalItemPage:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CanonicalItem"
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
    CreateCanonicalItemInput:
      type: object
      required: [legacyItemIds]
      properties:
        legacyItemIds:
          type: array
          minItems: 1
          items:
            type: string
        nomenclature:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        pvms:
          type: ["string", "null"]
        niv:
          type: ["string", "null"]
    VocabularyReview:
      type: object
      required: [id, importId, reviewType, identifier, title, detail, status, candidateRecords]
      properties:
        id:
          type: string
        importId:
          type: string
        reviewType:
          type: string
        identifier:
          type: ["string", "null"]
        title:
          type: string
        detail:
          type: string
        status:
          type: string
        decision:
          type: ["string", "null"]
        decisionNote:
          type: ["string", "null"]
        candidateRecords:
          type: array
          items:
            $ref: "#/components/schemas/LegacyItem"
    VocabularyDecisionInput:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [MERGE, KEEP_SEPARATE, CORRECT, RETIRE, CREATE_CANONICAL, INVESTIGATE]
        note:
          type: string
        canonicalItemIds:
          type: array
          items:
            type: string
        sourceColumnStart:
          type: integer
...