# lib/api-zod/src/generated/types/canonicalItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CanonicalItem {
  id: string;
  canonicalId: string;
  /** @nullable */
  nomenclature: string | null;
  /** @nullable */
  unit: string | null;
  /** @nullable */
  pvms: string | null;
  /** @nullable */
  niv: string | null;
  status: string;
  legacyRecordCount: number;
  departmentCount: number;
}