# lib/api-zod/src/generated/types/canonicalItemDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CanonicalItem } from './canonicalItem';
import type { CanonicalItemDetailVocabularyHistoryItem } from './canonicalItemDetailVocabularyHistoryItem';
import type { LegacyItem } from './legacyItem';

export type CanonicalItemDetail = CanonicalItem & {
  legacyRecords: LegacyItem[];
  vocabularyHistory: CanonicalItemDetailVocabularyHistoryItem[];
};