# lib/api-zod/src/generated/types/canonicalItemPage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CanonicalItem } from './canonicalItem';

export interface CanonicalItemPage {
  items: CanonicalItem[];
  page: number;
  pageSize: number;
  total: number;
}