# lib/api-zod/src/generated/types/createCanonicalItemInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCanonicalItemInput {
  /** @minItems 1 */
  legacyItemIds: string[];
  /** @nullable */
  nomenclature?: string | null;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  pvms?: string | null;
  /** @nullable */
  niv?: string | null;
}