# lib/api-zod/src/generated/types/vocabularyReview.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LegacyItem } from './legacyItem';

export interface VocabularyReview {
  id: string;
  importId: string;
  reviewType: string;
  /** @nullable */
  identifier: string | null;
  title: string;
  detail: string;
  status: string;
  /** @nullable */
  decision?: string | null;
  /** @nullable */
  decisionNote?: string | null;
  candidateRecords: LegacyItem[];
}